backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
google-cloud-storage==2.19.0
google-cloud-speech==2.28.0
google-generativeai==0.8.3
pydantic==2.10.4
